requirements.txt
algoliasearch-django>=3.0,<4.0
beautifulsoup4
celery
django>=4.0.0,<4.1.0
django-celery-beat
django-celery-results
djangorestframework
djangorestframework-simplejwt
matplotlib
notebook
pandas
pillow
python-decouple
pyyaml
redis
requests
seaborn
selenium
django-cors-headers